Cleanup scheduler

+++ scheduler/src/main/kotlin/co/nilin/opex/chainscan/scheduler/spi/ChainSyncSchedulerHandler.kt
@@ -6,5 +6,5 @@ import java.time.LocalDateTime
 interface ChainSyncSchedulerHandler {
     suspend fun fetchActiveSchedules(time: LocalDateTime): List<ChainSyncSchedule>
     suspend fun prepareScheduleForNextTry(syncSchedule: ChainSyncSchedule, success: Boolean)
-    suspend fun scheduleChain(chain: String, delaySeconds: Int, errorDelaySeconds: Int)
+    suspend fun scheduleChain(chain: String, delaySeconds: Long, errorDelaySeconds: Long)
 }
